// types/patient.ts
export interface Patient {

  id?: string;

  firstName: string;

  lastName: string;

  phone: string;

  email: string;

  dateOfBirth: string;

  gender: string;

  identificationType: string;

  identificationNumber: string;

  emergencyContactName: string;

  emergencyContactRelationship: string;

  emergencyContactPhone: string;

  bloodGroup: string;

  allergies: string;

  existingConditions: string;

  medicalNotes: string;

}